+++ analysis/src/Xform.ml
@@ -110,9 +110,20 @@ module IfThenElse = struct
     | Some newExpr ->
       let range = rangeOfLoc newExpr.pexp_loc in
       let newText = printExpr ~range newExpr in
+      let uri = Uri.fromPath path |> Uri.toString in
       let codeAction =
         CodeActions.make ~title:"Replace with switch" ~kind:RefactorRewrite
-          ~uri:path ~newText ~range
+          ~edit:
+            {
+              documentChanges =
+                [
+                  TextDocumentEdit
+                    {
+                      textDocument = {version = None; uri};
+                      edits = [{newText; range}];
+                    };
+                ];
+            }
       in
       codeActions := codeAction :: !codeActions
 end
@@ -174,9 +185,20 @@ module AddBracesToFn = struct
     | Some newStructureItem ->
       let range = rangeOfLoc newStructureItem.pstr_loc in
       let newText = printStructureItem ~range newStructureItem in
+      let uri = Uri.fromPath path |> Uri.toString in
       let codeAction =
         CodeActions.make ~title:"Add braces to function" ~kind:RefactorRewrite
-          ~uri:path ~newText ~range
+          ~edit:
+            {
+              documentChanges =
+                [
+                  TextDocumentEdit
+                    {
+                      textDocument = {version = None; uri};
+                      edits = [{newText; range}];
+                    };
+                ];
+            }
       in
       codeActions := codeAction :: !codeActions
 end
@@ -249,12 +271,142 @@ module AddTypeAnnotation = struct
               ( rangeOfLoc locItem.loc,
                 "(" ^ name ^ ": " ^ (typ |> Shared.typeToString) ^ ")" )
           in
+          let uri = Uri.fromPath path |> Uri.toString in
           let codeAction =
             CodeActions.make ~title:"Add type annotation" ~kind:RefactorRewrite
-              ~uri:path ~newText ~range
+              ~edit:
+                {
+                  documentChanges =
+                    [
+                      TextDocumentEdit
+                        {
+                          textDocument = {version = None; uri};
+                          edits = [{newText; range}];
+                        };
+                    ];
+                }
           in
           codeActions := codeAction :: !codeActions
         | _ -> ()))
+end
+
+module TypeToModule = struct
+  (* Convert type definition into its own module *)
+  let mkIterator ~pos ~result ~newTypeName =
+    let changeTypeDecl (typ : Parsetree.type_declaration) ~txt =
+      let manifest = typ.ptype_manifest in
+      Ast_helper.Type.mk
+        {txt; loc = typ.ptype_name.loc}
+        ~loc:typ.ptype_loc ~attrs:typ.ptype_attributes ~params:typ.ptype_params
+        ~cstrs:typ.ptype_cstrs ~kind:typ.ptype_kind ~priv:typ.ptype_private
+        ?manifest
+    in
+
+    let structure_item (iterator : Ast_iterator.iterator)
+        (si : Parsetree.structure_item) =
+      match si.pstr_desc with
+      | Pstr_type (Nonrecursive, firstypeDec :: rest)
+        when Loc.hasPos ~pos si.pstr_loc ->
+        let loc = si.pstr_loc in
+        let firsLetterOfModule =
+          String.get firstypeDec.ptype_name.txt 0
+          |> Char.uppercase_ascii |> String.make 1
+        in
+        let restName =
+          String.sub firstypeDec.ptype_name.txt 1
+            (String.length firstypeDec.ptype_name.txt - 1)
+        in
+        let modName = firsLetterOfModule ^ restName in
+        let newFistType = changeTypeDecl firstypeDec ~txt:newTypeName in
+        let restStrucTypes =
+          Ast_helper.Str.type_ ~loc Nonrecursive ([newFistType] @ rest)
+        in
+        let pmb_expr = Ast_helper.Mod.structure ~loc [restStrucTypes] in
+        let mod_expr =
+          Parsetree.Pstr_module
+            {
+              pmb_name = {txt = modName; loc};
+              pmb_expr;
+              pmb_attributes = [];
+              pmb_loc = loc;
+            }
+        in
+        let processTypeKind (typeDec : Parsetree.type_declaration) =
+          match typeDec.ptype_kind with
+          | Ptype_record [{pld_name = {loc; txt}}; _] -> Some (loc, txt)
+          | Ptype_variant [{pcd_name = {loc; txt}}; _] -> Some (loc, txt)
+          | _ -> None
+        in
+        let references =
+          let referencesInfo =
+            [firstypeDec] @ rest |> List.filter_map processTypeKind
+          in
+          [(firstypeDec.ptype_name.loc, newTypeName)] @ referencesInfo
+        in
+        result := Some (Ast_helper.Str.mk ~loc mod_expr, references, modName);
+        Ast_iterator.default_iterator.structure_item iterator si
+      | _ -> ()
+    in
+    {Ast_iterator.default_iterator with structure_item}
+
+  let xform ~path ~pos ~codeActions ~printStructureItem structure ~debug =
+    let result = ref None in
+    let newTypeName = "t" in
+    let iterator = mkIterator ~pos ~result ~newTypeName in
+    iterator.structure iterator structure;
+    match !result with
+    | None -> ()
+    | Some (newStructureItem, references, modName) ->
+      let range = rangeOfLoc newStructureItem.pstr_loc in
+      let newText = printStructureItem ~range newStructureItem in
+      let uri = Uri.fromPath path |> Uri.toString in
+      let typeToModuleEdit =
+        Protocol.TextDocumentEdit
+          {textDocument = {version = None; uri}; edits = [{newText; range}]}
+      in
+      (* Before apply code action find all references and rename with new name *)
+      let relatedChanges =
+        references
+        |> List.filter_map (fun (loc, txt) ->
+               let range = rangeOfLoc loc in
+               let newName = modName ^ "." ^ txt in
+               match
+                 Rename.command ~path
+                   ~pos:(range.start.line, range.start.character)
+                   ~newName ~debug
+               with
+               | Some workspaceEdit ->
+                 let result =
+                   workspaceEdit.documentChanges
+                   |> List.map
+                        (fun (documentChange : Protocol.documentChange) ->
+                          match documentChange with
+                          | TextDocumentEdit textEdit ->
+                            let textDocument = textEdit.textDocument in
+                            let edits =
+                              if textDocument.uri = uri then
+                                textEdit.edits
+                                |> List.filter
+                                     (fun (textEdit : Protocol.textEdit) ->
+                                       (* Ignore some text edit because it refers to the type that will be moved to its own module *)
+                                       textEdit.range <> range)
+                              else textEdit.edits
+                            in
+                            Protocol.TextDocumentEdit {textDocument; edits}
+                          | _ -> documentChange)
+                 in
+                 Some result
+               | None -> None)
+        |> List.flatten
+      in
+      let edit : Protocol.workspaceEdit =
+        {documentChanges = [typeToModuleEdit] @ relatedChanges}
+      in
+      let codeAction =
+        CodeActions.make ~title:"Move type definition into its own module"
+          ~kind:RefactorRewrite ~edit
+      in
+      codeActions := codeAction :: !codeActions
 end
 
 let indent n text =
@@ -310,5 +462,7 @@ let extractCodeActions ~path ~pos ~currentFile ~debug =
     AddTypeAnnotation.xform ~path ~pos ~full ~structure ~codeActions ~debug;
     IfThenElse.xform ~pos ~codeActions ~printExpr ~path structure;
     AddBracesToFn.xform ~pos ~codeActions ~path ~printStructureItem structure;
+    TypeToModule.xform ~path ~pos ~codeActions ~printStructureItem structure
+      ~debug;
     !codeActions
   | _ -> []